fix: tiny change

+++ src/code/comb/ball_box_ddl.hpp
@@ -10,7 +10,7 @@ template <class mint>
 inline mint ball_box_ddl(u32, u32 box, vec<u64> const& pows, Binom<mint> const& binom) {
   mint ans = 0;
   bool f = box & 1;
-  for (u64 i = 1; i <= box; ++i) {
+  for (u32 i = 1; i <= box; ++i) {
     f ? (ans += binom.mCn(box, i) * pows[i]) : (ans -= binom.mCn(box, i) * pows[i]);
     f ^= 1;
   }
